+++ dali/operators/reader/loader/CMakeLists.txt
@@ -51,7 +51,6 @@ endif()
 set(DALI_OPERATOR_TEST_SRCS ${DALI_OPERATOR_TEST_SRCS}
   "${CMAKE_CURRENT_SOURCE_DIR}/loader_test.cc"
   "${CMAKE_CURRENT_SOURCE_DIR}/sequence_loader_test.cc"
-  "${CMAKE_CURRENT_SOURCE_DIR}/numpy_loader_test.cc"
   "${CMAKE_CURRENT_SOURCE_DIR}/filesystem_test.cc")
 
 if (BUILD_LIBSND)
