+++ dali/operators/reader/numpy_reader_op.h
@@ -15,10 +15,6 @@
 #ifndef DALI_OPERATORS_READER_NUMPY_READER_OP_H_
 #define DALI_OPERATORS_READER_NUMPY_READER_OP_H_
 
-#define NUMPY_ALLOWED_TYPES \
-  (bool, uint8_t, uint16_t, uint32_t, uint64_t, int8_t, int16_t, int32_t, int64_t, float, float16, \
-  double)
-
 #define NUMPY_ALLOWED_DIMS (0, 1, 2, 3, 4, 5, 6)
 
 #include <string>
